Add first draft
This is the first template for the app with some functionality. ATM working with just a local database. User registering, login etc. is implemented as it was in course material (not copy-paste, but pretty much typed line for line) - might look into it, if some changes are needed. App now has login functionality and users are able to add new recipes and view recipes. But users are f.e. able to only insert three different ingredients for their recipe, so everything is really barebones and just a template.